+++ src/main/groovy/com/monits/gradle/sca/config/PmdConfigurator.groovy
@@ -19,6 +19,7 @@ import com.monits.gradle.sca.StaticCodeAnalysisExtension
 import com.monits.gradle.sca.ToolVersions
 import groovy.transform.CompileStatic
 import groovy.transform.TypeCheckingMode
+import org.gradle.api.InvalidUserDataException
 import org.gradle.api.NamedDomainObjectContainer
 import org.gradle.api.Namer
 import org.gradle.api.Project
@@ -34,19 +35,24 @@ import org.gradle.api.tasks.compile.JavaCompile
 import org.gradle.util.GUtil
 import org.gradle.util.GradleVersion
 
+import java.util.regex.Matcher
+
 /**
  * A configurator for PMD tasks.
 */
 @CompileStatic
-class PmdConfigurator implements AnalysisConfigurator, ClasspathAware {
+class PmdConfigurator extends AbstractRemoteConfigLocator implements AnalysisConfigurator, ClasspathAware {
     private final static GradleVersion GRADLE_VERSION_PMD_CLASSPATH_SUPPORT = GradleVersion.version('2.8')
     private final static String PMD = 'pmd'
+
+    final String pluginName = PMD
 
     @Override
     void applyConfig(final Project project, final StaticCodeAnalysisExtension extension) {
         setupPlugin(project, extension)
 
         SourceSetContainer sourceSets = project.convention.getPlugin(JavaPluginConvention).sourceSets
+        //noinspection GroovyMissingReturnStatement
         setupTasksPerSourceSet(project, extension, sourceSets) { Pmd pmdTask, SourceSet sourceSet ->
             boolean supportsClasspath = GRADLE_VERSION_PMD_CLASSPATH_SUPPORT <= GradleVersion.current()
 
@@ -62,6 +68,7 @@ class PmdConfigurator implements AnalysisConfigurator, ClasspathAware {
     void applyAndroidConfig(final Project project, final StaticCodeAnalysisExtension extension) {
         setupPlugin(project, extension)
 
+        //noinspection GroovyAssignabilityCheck
         setupTasksPerSourceSet(project, extension, project.android.sourceSets) { Pmd pmdTask, sourceSet ->
             /*
              * Android doesn't expose name of the task compiling the sourceset, and names vary
@@ -95,8 +102,8 @@ class PmdConfigurator implements AnalysisConfigurator, ClasspathAware {
         }
     }
 
-    @SuppressWarnings('UnnecessaryGetter')
-    private static void setupTasksPerSourceSet(final Project project, final StaticCodeAnalysisExtension extension,
+    @SuppressWarnings(['UnnecessaryGetter', 'DuplicateStringLiteral'])
+    private void setupTasksPerSourceSet(final Project project, final StaticCodeAnalysisExtension extension,
                                                final NamedDomainObjectContainer<?> sourceSets,
                                                final Closure<?> configuration = null) {
         // Create a phony pmd task that just executes all real pmd tasks
@@ -106,11 +113,51 @@ class PmdConfigurator implements AnalysisConfigurator, ClasspathAware {
             String sourceSetName = namer.determineName(sourceSet)
             RulesConfig config = extension.sourceSetConfig.maybeCreate(sourceSetName)
 
-            Task pmdTask = getOrCreateTask(project, generateTaskName(sourceSetName)) { Pmd it ->
-                // most defaults are good enough
+            List<Task> downloadTasks = []
+            List<String> rulesets = []
+
+            for (String ruleset : config.getPmdRules()) {
+                boolean remoteLocation = isRemoteLocation(ruleset)
+                File configSource
+                if (remoteLocation) {
+                    Matcher filenameMatcher = ruleset =~ /\/([^\/]*)\.[^.]*$/
+                    filenameMatcher.find()
+                    String filename = filenameMatcher.group(1).replaceAll '[^a-zA-Z0-9]', ' '
+                    String downloadTaskName
+
+                    /*
+                     * Names may have collisions if the same file is downloaded from different domains..
+                     * add suffixes as needed
+                     */
+                    int attempts = 0
+                    String suffix = ''
+                    while (configSource == null) {
+                        try {
+                            downloadTaskName = generateTaskName('downloadPmdXml', sourceSetName, filename) + suffix
+                            configSource = makeDownloadFileTask(project, ruleset,
+                                String.format('pmd-%s-%s.xml', sourceSetName, filename.replaceAll(/ /, '-') + suffix),
+                                downloadTaskName)
+                        } catch (InvalidUserDataException ignored) {
+                            attempts++
+                            suffix = attempts
+                        }
+                    }
+                    downloadTasks.add(project.tasks.findByName(downloadTaskName))
+                } else {
+                    configSource = project.file(ruleset)
+                }
 
                 // PMD doesn't play ball with relative paths nor file:// URIs
-                it.ruleSets = config.getPmdRules().collect { it =~ /https?:\/\// ? it : project.file(it).absolutePath }
+                rulesets.add(configSource.absolutePath)
+            }
+
+            Task pmdTask = getOrCreateTask(project, generateTaskName(sourceSetName)) { Pmd it ->
+                // most defaults are good enough
+                if (!downloadTasks.empty) {
+                    it.dependsOn downloadTasks
+                }
+
+                it.ruleSets = rulesets
 
                 it.reports { PmdReports r ->
                     r.with {
@@ -144,7 +191,8 @@ class PmdConfigurator implements AnalysisConfigurator, ClasspathAware {
         pmdTask.configure closure
     }
 
-    private static String generateTaskName(final String sourceSetName) {
-        GUtil.toLowerCamelCase(String.format('%s %s', PMD, sourceSetName))
+    private static String generateTaskName(final String taskName = PMD, final String sourceSetName,
+                                           final String classifier = '') {
+        GUtil.toLowerCamelCase(String.format('%s %s %s', taskName, sourceSetName, classifier))
     }
 }